==> Material_apoyo/01_ConceptosPython.ipynb <==
# **Ciencia de Datos**
---
**Conceptos de Python**
* Python es un lenguaje de programación de alto nivel, interpretado y de propósito general.
* Python se caracteriza por ser fácil de leer y escribir, lo que lo hace ideal para principiantes en la programación, pero también es utilizado por expertos en una amplia gama de aplicaciones, desde la ciencia de datos y la inteligencia artificial hasta el desarrollo web y la automatización de tareas.
* Es un lenguaje multiplataforma, lo que significa que puede ser ejecutado en sistemas operativos como Windows, Linux y MacOS
* Python es un lenguaje interpretado, lo que significa que no es necesario compilar el código antes de ejecutarlo, lo que agiliza el proceso de desarrollo y lo hace ideal para prototipos rápidos.

**Variables**

En Python, una variable es un nombre que se utiliza para referirse a un valor almacenado en la memoria de la computadora. Las variables en Python se pueden definir y utilizar de la siguiente manera:
* Para definir una variable, se utiliza el signo igual (=) y se asigna un valor a un nombre específico.
* En Python, las variables pueden contener diferentes tipos de datos, como números enteros, números de punto flotante, cadenas de texto y booleanos (True o False). El tipo de datos se infiere automáticamente en función del valor asignado a la variable.
* Las variables pueden cambiar de valor durante la ejecución del programa. 

# Definición de variables y tipos de datos
nombre = "Juan"
edad = 25
estado = True

# Cambios de valor
edad = 30
y = 10
x = y

# Conversión de valores
x = str(3)
y = int(3)
z = float(3)

# Asignación de valores
uno, dos, tres = "Lunes", "Martes", "Miercoles"
j = k = h = "Verde"

# **Entradas y Salidas de usuario**

En Python, el término "entrada" (input) se refiere a cualquier información que el usuario proporciona al programa, mientras que la "salida" (output) se refiere a cualquier información que el programa muestra al usuario. Python ofrece varias formas de leer la entrada y mostrar la salida.

**Entrada de usuario (input)**: 
La forma más común de obtener la entrada de usuario en Python es mediante la función input(). Esta función lee una línea de texto que el usuario escribe en la consola y devuelve esa línea de texto como una cadena de caracteres.

nombre = input("Por favor ingrese su nombre: ")
edad = input("Por favor ingrese su edad: ")
peso = float(input("Por favor ingrese su peso: "))

**Salida de texto (output)**: 
La forma más simple de mostrar texto en Python es mediante la función print(). Esta función toma uno o más argumentos y los muestra en la consola separados por espacios. 

print("Hola, mundo!")
print("Hola, " + nombre + "!")
print("Hola, ", nombre ,"! y tienes, ", edad, "años y pesas ", peso , "kg")
print("Hola, me llamo {} y tengo {} años.".format(nombre, edad))
print(f"Hola, me llamo {nombre} y tengo {edad} años.")

# **Operadores**
En Python, los operadores son símbolos que se utilizan para realizar operaciones sobre valores y variables. Python cuenta con una amplia variedad de operadores, que se pueden clasificar en diferentes categorías:

**Operadores aritméticos**: 
Los operadores aritméticos se utilizan para realizar operaciones matemáticas básicas, tales como sumar, restar, multiplicar, dividir y calcular el resto de una división. Los operadores aritméticos en Python son los siguientes: **+** (suma), **-** (resta), * (multiplicación), **/** (división), **%** (resto de la división), ** (potencia), **//** (división entera).

# **Estructura secuencial**

La estructura secuencial en Python se refiere a la ejecución de instrucciones en un orden específico, de manera que una instrucción se ejecuta después de la otra en secuencia.

La estructura secuencial es la forma más básica de programación y se utiliza para resolver problemas sencillos que no requieren ramificación ni iteración. Las instrucciones en la estructura secuencial se ejecutan en el orden en que aparecen en el programa, de arriba hacia abajo.

# Determinar el area y perimetro de un circulo
# Area = Pi * Radio2 y Perimetro = 2 * Pi * r

# definicion de Variables
pi = 3.141516

# Entrada
radio = float(input("Por favor ingrese radio: "))

# Proceso
area = pi * (radio**2)
perimetro = 2 * pi * radio

# Salida
print(f"Area es: {round(area,2)} y el perimetro es: {round(perimetro,2)} del Circulo.")

**Operadores de comparación**

Los operadores de comparación se utilizan para comparar dos valores y devolver un valor booleano (True o False) que indica si la comparación es verdadera o falsa. Los operadores de comparación en Python son los siguientes:
**==** (igual a), **!=** (diferente de), **<** (menor que), **>** (mayor que), **<=** (menor o igual que), **>=** (mayor o igual que).


**Operadores lógicos**

Los operadores lógicos se utilizan para combinar valores booleanos y devolver otro valor booleano. Los operadores lógicos en Python son los siguientes: **and** (y), **or** (o), **not** (no).

# **Estructura condicional**
La estructura condicional en Python se utiliza para ejecutar diferentes bloques de código según si se cumple o no una condición (**if-else**). En esta estructura, la **condicion** es una expresión booleana que se evalúa como verdadera o falsa. Si la condicion es **verdadera**, se ejecuta el bloque de código que se encuentra debajo del **if**, y si es **falsa**, se ejecuta el bloque de código que se encuentra debajo del **else**.


Además de la estructura if-else, en Python también existe la estructura **if-elif-else**, que se utiliza para evaluar múltiples condiciones. En esta estructura, se evalúan las condiciones en orden, y se ejecuta el bloque de código correspondiente al primer caso que se cumpla. Si ninguna de las condiciones se cumple, se ejecuta el bloque de código que se encuentra debajo del else.

# Determinar el un usuario cumple los requisitos para prestar el servicio militar
# Requisitos = Hombre, mayor o igual a 18 años

# Entrada
genero = str(input("Por favor ingrese genero (M o F): ")).upper()
edad = int(input("Por favor ingrese su edad: "))

# Proceso - salida
if(genero == "M" and edad >= 18):
  print("Cumple con los requisitos para el Servicio Militar")
else:
  print("NO cumple con los requisitos para el Servicio Militar")

# Determinar el un numero entero es positivo, negativo o cero
numero = int(input("Ingrese un número: "))

if numero > 0:
    print("El número ingresado es positivo")
elif numero < 0:
    print("El número ingresado es negativo")
else:
    print("El número ingresado es cero")

# **Estructuras repetitivas**
Las estructuras repetitivas, también conocidas como bucles, se utilizan en Python para repetir un bloque de código varias veces. En Python, existen dos tipos de bucles: **while** y **for**.

* La **estructura while** en Python se utiliza para repetir un bloque de código mientras se cumpla una condición.

* La **estructura for** en Python se utiliza para repetir un bloque de código para cada elemento de una secuencia.

# Estructura repetitiva while en Python
numero = 1

while numero <= 10:
    print(numero)
    numero += 1

# Estructura repetitiva for en Python
texto = "Hola, mundo!"

for caracter in texto:
    print(caracter)

# **Estructuras de datos**
En Python, un **diccionario** es una estructura de datos que se utiliza para almacenar datos en pares de **clave-valor**. Cada clave es única en el diccionario y se utiliza para acceder al valor correspondiente. Cada clave se escribe entre comillas y se separa del valor correspondiente por dos puntos. Los pares clave-valor se separan por comas, y todo el diccionario se encierra entre llaves.

diccionario = {"clave1": valor1, "clave2": valor2, ...}

# Crear un diccionario en Python
personas = {"Juan": 25, "Maria": 30, "Pedro": 35, "Alvaro":34, "Daniel": 33, "Ediccson":34}

# Acceder al valor de una clave en un diccionario
edad = personas["Maria"]
print("La edad de Maria es:", edad)

# Añadir un elemento al un diccionario
personas["Luis"] = 28

# Modificar un elemento al un diccionario
personas["Juan"] = 40

# Eliminar un elemento al un diccionario
personas.pop("Ediccson")

# Recorrer un diccionario con un bucle for
for nombre, edad in personas.items():
    print(nombre, "tiene:", edad, "años")


# Adicionales
print(len(personas)) #Numero de elementos
print(type(personas)) #Tipo de variable

# **Funciones**
En Python, una función es un bloque de código que realiza una tarea específica y que puede ser reutilizado en diferentes partes de un programa. Una función se define utilizando la palabra clave def seguida del nombre de la función y los parámetros de entrada entre paréntesis.


def nombre_de_la_funcion(parametro1, parametro2, ...):
    # Cuerpo de la función
    # Realiza una tarea específica utilizando los parámetros de entrada
    # y posiblemente variables internas
    
    return resultado

Existen diferentes tipos de funciones que se pueden utilizar según la necesidad del programa. A continuación se presentan algunos de los tipos de funciones más comunes:

1.   **Funciones sin parámetros:** Son aquellas que no requieren de ningún parámetro de entrada para realizar su tarea.

def saludar():
    print("Hola, ¿cómo estás?")

2. **Funciones con parámetros:** Son aquellas que reciben uno o más parámetros de entrada que son necesarios para realizar su tarea.

def suma(a, b):
    resultado = a + b
    print(resultado)

3. **Funciones con valor de retorno:** Son aquellas que devuelven un valor como resultado de su ejecución.

def multiplicar(a, b):
    resultado = a * b
    return resultado
