==> Material_apoyo/06_Visualizacion.ipynb <==
# **Visualización de datos estadísticos**

x

**Visualización de datos cuantitativos:** Una de las primeras cosas que hay que hacer cuando nos enfrentamos a datos cuantitativos (o a cualquier tipo de datos) es hacer algún tipo de representación visual para comprender la estructura de los datos. La visualización más común para datos continuos es el histograma, el cual representa la distribución de los valores

Un histograma se construye decidiendo en primer lugar un número de intervalos (o *bins*), que dividen la línea real en un conjunto de cajas en las que caen los valores de los datos. Luego, se dibujan barras verticales sobre los intervalos con una altura proporcional al número de observaciones que caen en el intervalo. 

Podemos construir un histograma con matplotlib utilizando la función `plt.hist`, la cual toma como argumentos la variable a graficar y el número de bins (entre otros)
